ponto_app_final_v5/requirements.txt
Flask==2.3.3
Flask_SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
gunicorn==20.1.0
psycopg2-binary==2.9.7
reportlab==4.0.4
openpyxl==3.1.2
itsdangerous==2.1.2
Flask-Mail==0.10.0
supabase==2.18.1